fix: keep workspace transfer type out of budgeted file
